+++ Kokoro/docs/release-notes.md
@@ -6,6 +6,7 @@
 
 * Exhaust no longer appears as a trait for Limited 1 cards.
 * Infinite no longer appears as a trait for Finite 2+ cards.
+* Improved the tooltips of Limited and Finite.
 * Fixed "Times Played This Turn" not resetting under some circumstances.
 
 ## 2.12.2